requirements.txt

Flask==1.1.1
pytesseract==0.3.10
Pillow==8.0.1
